+++ src/Storage/Filesystem.php
@@ -15,6 +15,7 @@
 namespace Phossa2\Storage;
 
 use Phossa2\Shared\Base\ObjectAbstract;
+use Phossa2\Storage\Driver\LocalDriver;
 use Phossa2\Storage\Traits\DriverAwareTrait;
 use Phossa2\Storage\Interfaces\DriverInterface;
 use Phossa2\Storage\Traits\PermissionAwareTrait;
@@ -30,8 +31,9 @@ use Phossa2\Storage\Interfaces\PermissionAwareInterface;
  * @author  Hong Zhang <phossa@126.com>
  * @see     ObjectAbstract
  * @see     FilesystemInterface
- * @version 2.0.0
+ * @version 2.1.0
  * @since   2.0.0 added
+ * @since   2.1.0 modified `__construct` to accept string as first param
  */
 class Filesystem extends ObjectAbstract implements FilesystemInterface
 {
@@ -40,17 +42,23 @@ class Filesystem extends ObjectAbstract implements FilesystemInterface
     /**
      * Set the driver and its global permissions
      *
-     * @param  DriverInterface $driver
+     * @param  DriverInterface|string $driverOrPath driver or local path
      * @param  int $permissions filesystem permissions
      * @access public
+     * @since  2.1.0 changed first param
      * @api
      */
     public function __construct(
-        DriverInterface $driver,
+        $driverOrPath,
         /*# int */ $permissions = PermissionAwareInterface::PERM_ALL
     ) {
+        // path provided, init a LocalDriver
+        if (is_string($driverOrPath)) {
+            $driverOrPath = new LocalDriver($driverOrPath);
+        }
+
         // set underlying driver
-        $this->setDriver($driver);
+        $this->setDriver($driverOrPath);
 
         // set permissions for THIS filesystem
         $this->setPermissions($permissions);